backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
tavily-python
chromadb
sentence-transformers
langgraph
langchain
pypdf
langchain-community
tiktoken
redis
